Refactor plugin to allow a whitelabel fork and fix a few bugs while at it
Always initialize instance_form_fields to avoid notices from $this->get_instance_option('shipping_method')

Apply namespaces, move things around and fix notices

Use autoloading, restructure plugin

Reorder things and parse all values

Hardcode text domain again

Reasoning in #84

Flatten structure and start moving translatable strings to a class

Disable autoloader as it's classname conflicts with whitelabel plugin, recreate tests (untested) and make every class overloadable

Rename assets

Convert rest of shipping method class

Change tests to use existing instances instead of creating new ones

Add scripts for locally running tests

Download WooCommerce in install-wp-tests.sh

Bumb php version

Delete duplicate files

Move away from index.php

Actually allow overloading

Get rid of singleton

Allow changing setuppage dynamically

Remove static $core

Remove slashes and hardcoded name

Change namespace

Remove "wc_pakettikauppa" from code to make str replacing easier

Replace most instances of wc_pakettikauppa with variable

Replace most instances of pakettikauppa with str_replace

TODO: Bumb plugin version to 3, and replace str_replace("wc_", "",
$this->core->prefix) with just $this->core->prefix. It will break
existing sites with unprocessed orders.

Replace woocommerce-pakettikauppa with woo-pakettikauppa

Replace wc-pakettikauppa with woo-pakettikauppa

Change metabox id to allow metaboxes from both plugins

Allow preventing load

Load text class earlier

Hardcode and add a reminder

Add __DIR__

Remove extra class instance

Replace with variable

Add a few words on architecture

Load shipment only once

Fix creation in whitelabel

Fix rest of methods

Missed some words
