+++ .travis.yml
@@ -8,15 +8,16 @@ env:
   - WP_VERSION=latest
 
 php:
-  - '7.2'
   - '7.3'
+  - '7.2'
+  - '7.1'
   - nightly
 
 matrix:
-  include:
+  # include:
     # Arbitrary PHP version to run the sniffs against.
-    - php: '5.6'
-      env: WP_VERSION=4.9.9 WP_MULTISITE=1 SNIFF=1
+    # - php: '5.6'
+      # env: WP_VERSION=4.9.9 WP_MULTISITE=1 SNIFF=1
   allow_failures:
   - php: nightly  # Old phpunit not compatible with PHP 8.0.0 yet
