+++ README.md
@@ -74,3 +74,27 @@ This project has a `.travis-ci.yml` definition which can be used by anybody. Jus
 ![Travis CI: Activate](docs/travis-ci-repo-active.png)
 5. Initially the page `https://travis-ci.org/<username>/woo-pakettikauppa` will be empty, but as soon as you push your first commit Travis-CI will run the first built. You can also manually trigger a build to get it started.
 ![Travis CI: Trigger manual build](docs/travis-ci-manual-trigger.png)
+
+## Running tests locally
+
+You can run the tests locally with by using the install-tests.sh & run-tests.sh. Note that these scripts are **destructive**, so you should probably run them in a virtual environment such as Vagrant or Docker. Assuming you're using [Seravo/wordpress](https://github.com/Seravo/wordpress):
+
+```
+docker-compose exec wordpress bash
+cd /data/wordpress/htdocs/wp-content/plugins/woo-pakettikauppa/bin
+
+# do not run ./install-wp-tests.sh yourself
+./install-local-tests.sh
+./run-local-tests.sh
+```
+
+You may need to chmod the scripts before using: `chmod +x install-tests.sh`
+
+## Architecture
+The plugin is written so that it can be easily forked to a whitelabel version, containing only a subset of shipping providers and a custom branding. This is achieved by using an abstract class as a factory for the plugin component. Each component can be replaced by simply overloading a method.
+
+The `Woo_Pakettikauppa_Core\Core` class is responsible for loading all components, and it serves as a bridge between the components. The instance is passed to most subclasses of the plugin, such as `Woo_Pakettikauppa_Core\Admin`, which keep a reference to it under `$this->core`.
+
+The subclasses are responsible for doing the actual heavy lifting. `Admin` handles operations in wp-admin, `Frontend` handles frontend, `Shipment` is used for creating shipment labels, etc.
+
+`Text` is bit of a special case, it's only real purpose is to serve as a key-value store for strings that require translation. Currently the plugin has __ & _e calls scattered all over it, in the future, all of them should live in `Text`. The advantage of doing so is being able to replace any text in the whitelabel forks.
