+++ assets/css/admin-setup.css
