--- assets/css/wc-pakettikauppa/admin-setup.css
@@ -1,141 +1,0 @@
-html, body {
-  height: auto !important;
-  min-height: 100%;
-}
-.wcpk-setup-body {
-  background-repeat: no-repeat;
-  background-position: center center;
-  background-size: cover;
-}
-.wcpk-setup {
-  margin: 70px auto 40px;
-  text-align: center;
-  max-width: 700px;
-  padding: 0;
-}
-.wcpk-setup .wcpk-setup-wizard-content .wcpk-setup-actions {
-  text-align: center;
-}
-.wcpk-setup .wcpk-setup-steps {
-  list-style: none outside;
-  display: inline-block;
-  overflow: hidden;
-  text-align: center;
-  margin: 0;
-  padding: 0 0 10px;
-}
-.wcpk-setup .wcpk-setup-steps li {
-  float: left;
-  text-align: center;
-  color: #ccc;
-  padding: 0 1em .8em;
-  position: relative;
-  border-bottom: 4px solid rgba(204, 204, 204, 0.2);
-}
-.wcpk-setup .wcpk-setup-steps li::before {
-  content: "";
-  border: 4px solid rgba(204, 204, 204, 0.5);
-  border-radius: 100%;
-  width: 4px;
-  height: 4px;
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  margin-left: -6px;
-  margin-bottom: -8px;
-}
-.wcpk-setup .wcpk-setup-steps li.completed {
-  border-bottom: 4px solid #fff;
-  color: #fff;
-}
-.wcpk-setup .wcpk-setup-steps li.completed::before {
-  border: 4px solid #fff;
-}
-.wcpk-setup .wcpk-setup-steps li.active {
-  float: left;
-  text-align: center;
-  border-bottom: 4px solid #fff;
-  font-weight: 800;
-  color: #fff;
-}
-.wcpk-setup .wcpk-setup-steps li.active::before {
-  border: 4px solid #fff;
-}
-.wcpk-setup .wcpk-setup-settings-wrapper form > label {
-  margin-bottom: .5em;
-  margin-top: 1.2em;
-  display: inline-block;
-  font-weight: 700;
-}
-.wcpk-setup .wcpk-setup-settings-wrapper fieldset select,
-.wcpk-setup .wcpk-setup-settings-wrapper fieldset input {
-  width: 100%;
-}
-.wcpk-setup .wcpk-setup-settings-wrapper fieldset input[type="checkbox"] {
-  width: auto;
-}
-.wcpk-setup .wcpk-shipping-setup h3 {
-  font-size: 1.1em;
-}
-.wcpk-setup .wcpk-shipping-setup h4 {
-  text-decoration: underline;
-}
-.wcpk-setup .wcpk-shipping-setup table {
-  width: 100%;
-  margin-bottom: 1em;
-}
-.wcpk-setup .wcpk-shipping-setup p {
-  margin: .2em;
-}
-.wcpk-setup .wcpk-shipping-setup table table tr {
-  box-shadow: 0 0 1px rgba(0, 0, 0, .1);
-  padding: .2em;
-}
-.wcpk-setup .wcpk-shipping-setup tr th {
-  padding: .5em;
-  display: inline-block;
-}
-.wcpk-setup .wcpk-shipping-setup tr td {
-  padding: .5em;
-}
-.wcpk-setup .wcpk-shipping-setup tr td:first-of-type {
-  display: block;
-  padding: .2em 0 1em 1em;
-}
-.wcpk-setup .wcpk-setup-steps a,
-.wcpk-setup .wcpk-setup-footer-links a {
-  color: #fff;
-  text-decoration: none;
-}
-.wcpk-setup .wcpk-setup-steps a:hover,
-.wcpk-setup .wcpk-setup-steps a:active,
-.wcpk-setup .wcpk-setup-footer-links a:hover,
-.wcpk-setup .wcpk-setup-footer-links a:active {
-  color: #fff;
-  text-decoration: underline;
-}
-.wcpk-setup .screen-reader-text span {
-  clip-path: inset(50%);
-}
-.wcpk-setup #pakettikauppa-logo {
-  text-align: center;
-}
-.wcpk-setup #pakettikauppa-logo a img {
-  max-width: 30%;
-}
-.wcpk-setup .wcpk-setup-wizard-content {
-  border-radius: 3px;
-  box-shadow: 3px 3px 9px rgba(0, 0, 0, .2);
-  padding: 2em;
-  margin-bottom: 20px;
-  background: #f1f1f1;
-  text-align: left;
-  overflow: hidden;
-  zoom: 1;
-}
-.wcpk-setup .wcpk-setup-wizard-content fieldset {
-  border: none;
-}
-.wcpk-setup .wcpk-setup-wizard-content .wcpk-setup-actions {
-  margin-top: 1em;
-}
