--- assets/css/wc-pakettikauppa/admin.css
@@ -1,44 +1,0 @@
-p.pakettikauppa-shipment {
-  border-bottom: 1px solid #dddddd;
-  padding-left: 12px;
-  padding-right: 12px;
-  margin-left: -12px;
-  margin-right: -12px;
-  margin-bottom: 10px;
-  padding-bottom: 15px;
-}
-
-.pakettikauppa-services fieldset h4 {
-  margin: 10px 0 5px;
-}
-
-.pakettikauppa-services div#wc-pakettikauppa-pickup-points-wrapper {
-  width: 95%;
-}
-
-.wc_pakettikauppa_pickup_point_id {
-  width: 100%;
-}
-
-.pakettikauppa-services div#wc-pakettikauppa-cod-amount-wrapper,
-.pakettikauppa-services div#wc-pakettikauppa-cod-reference-wrapper,
-.pakettikauppa-services div#wc-pakettikauppa-pickup-points-wrapper
-{
-  display: none;
-}
-
-.button.wc-pakettikauppa-delete-button {
-    color: #a00 !important;
-}
-
-.wc-pakettikauppa-metabox-error {
-  color: #a00;
-}
-
-.widefat .column-wc_actions a.pakettikauppa_create_shipping_label::after {
-  content: '\f491';
-}
-
-select.pakettikauppa_metabox_values {
-  width: 100%;
-}
