--- assets/css/wc-pakettikauppa/frontend.css
@@ -1,3 +1,0 @@
-/**
-* Add frontend style to this file.
-*/
